Refactor library create modals layout
