+++ src/apps/library/create/creature/index.ts
@@ -1,0 +1,2 @@
+// src/apps/library/create/creature/index.ts
+export { CreateCreatureModal } from "./modal";
