+++ src/apps/library/create/spell/index.ts
@@ -1,0 +1,2 @@
+// src/apps/library/create/spell/index.ts
+export { CreateSpellModal } from "./modal";
